refactored: created values.ts in utilities
